#7126 move some functions to widgetHelpers, remove widgetHelper #7126 add missing props to WidgetRenderer
# Conflicts:
#	frontend/src/components/widget/RawWidget.js
